requirements.txt
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.4
arrow==0.17.0
astroid==2.4.2
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
binaryornot==0.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.3
connexion==2.7.0
cookiecutter==1.7.2
Creoleparser==0.7.5
defusedxml==0.6.0
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.1
env-file==0.0.0
fastapi==0.61.1
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
flask-blueprint==1.3.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-OpenID==1.2.5
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.3
from==0.1.0
Genshi==0.7.4
h11==0.11.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.24.0
mccabe==0.6.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
openapi-spec-validator==0.2.9
pipfile==0.0.2
poyo==0.5.0
protobuf==3.13.0
pycodestyle==2.6.0
pydantic==1.6.1
pyflakes==2.2.0
Pygments==2.7.2
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
PyMySQL==0.10.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-openid==2.2.5
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.6
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.12.1
values==0.0.0
virtualenv==20.0.31
Werkzeug==1.0.1
Whoosh==2.7.4
wrapt==1.12.1
WTForms==2.3.3
yapf==0.30.0
zipp==3.2.0
